pyproject.toml
[project]
name = "servicedeskplus-mcp"
version = "0.1.0"
description = "MCP server for ServiceDesk Plus"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.9.0",
]

[project.scripts]
servicedeskplus-mcp = "main:run"

[tool.setuptools]
py-modules = ["main", "sdp_client", "config", "admin_client"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
